backend set up finished but may have errors between frontend and backend data vars using Items and CaseStudies

+++ backend/src/items/items.router.ts
@@ -1,3 +1,4 @@
+//backend/src/items/items.router.ts
 /**
  * Required External Modules and Interfaces
  */
@@ -6,8 +7,8 @@ import * as ItemService from "./items.service";
 import { BaseItem, Item } from "./item.interface";
 
 import { checkJwt } from "../middleware/authz.middleware";
-
-
+import { checkPermissions } from "../middleware/permissions.middleware";
+import { ItemPermission } from "./item-permission";
 /**
  * Router Definition
  */
@@ -28,8 +29,8 @@ itemsRouter.get("/", async (req: Request, res: Response) => {
   }
 });
 // GET items/:id
-itemsRouter.get("/:id", async (req: Request, res: Response) => {
-  const id: number = parseInt(req.params.id, 10);
+itemsRouter.get("/:useCaseId", async (req: Request, res: Response) => {
+  const id: number = parseInt(req.params.useCaseId, 10);
 
   try {
     const item: Item = await ItemService.find(id);
@@ -43,55 +44,68 @@ itemsRouter.get("/:id", async (req: Request, res: Response) => {
     res.status(500).send(e.message);
   }
 });
-// ✨ New! Mount authorization middleware
-itemsRouter.use(checkJwt); // 👈 👀
+// Mount authorization middleware
+
+itemsRouter.use(checkJwt);
 
 // POST items
 
-itemsRouter.post("/", async (req: Request, res: Response) => {
-  try {
-    const item: BaseItem = req.body;
+itemsRouter.post(
+  "/",
+  checkPermissions(ItemPermission.CreateItems),
+  async (req: Request, res: Response) => {
+    try {
+      const item: BaseItem = req.body;
 
-    const newItem = await ItemService.create(item);
+      const newItem = await ItemService.create(item);
 
-    res.status(201).json(newItem);
-  } catch (e:any) {
-    res.status(500).send(e.message);
+      res.status(201).json(newItem);
+    } catch (e:any) {
+      res.status(500).send(e.message);
+    }
   }
-});
+);
 
 // PUT items/:id
 
-itemsRouter.put("/:id", async (req: Request, res: Response) => {
-  const id: number = parseInt(req.params.id, 10);
+itemsRouter.put(
+  "/:useCaseId",
+  checkPermissions(ItemPermission.UpdateItems),
+  async (req: Request, res: Response) => {
+    const id: number = parseInt(req.params.useCaseId, 10);
 
-  try {
-    const itemUpdate: Item = req.body;
+    try {
+      const itemUpdate: Item = req.body;
 
-    const existingItem: Item = await ItemService.find(id);
+      const existingItem: Item = await ItemService.find(id);
 
-    if (existingItem) {
-      const updatedItem = await ItemService.update(id, itemUpdate);
-      return res.status(200).json(updatedItem);
+      if (existingItem) {
+        const updatedItem = await ItemService.update(id, itemUpdate);
+        return res.status(200).json(updatedItem);
+      }
+
+      const newItem = await ItemService.create(itemUpdate);
+
+      res.status(201).json(newItem);
+    } catch (e:any) {
+      res.status(500).send(e.message);
     }
-
-    const newItem = await ItemService.create(itemUpdate);
-
-    res.status(201).json(newItem);
-  } catch (e:any) {
-    res.status(500).send(e.message);
   }
-});
+);
 
 // DELETE items/:id
 
-itemsRouter.delete("/:id", async (req: Request, res: Response) => {
-  try {
-    const id: number = parseInt(req.params.id, 10);
-    await ItemService.remove(id);
+itemsRouter.delete(
+  "/:id",
+  checkPermissions(ItemPermission.DeleteItems),
+  async (req: Request, res: Response) => {
+    try {
+      const id: number = parseInt(req.params.id, 10);
+      await ItemService.remove(id);
 
-    res.sendStatus(204);
-  } catch (e:any) {
-    res.status(500).send(e.message);
+      res.sendStatus(204);
+    } catch (e:any) {
+      res.status(500).send(e.message);
+    }
   }
-});
+);
